+++ src/main/java/dev/gacbl/logicore/blocks/research_station/ResearchStationLootTableProvider.java
@@ -1,0 +1,26 @@
+package dev.gacbl.logicore.blocks.research_station;
+
+import net.minecraft.core.Holder;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.loot.BlockLootSubProvider;
+import net.minecraft.world.flag.FeatureFlags;
+import net.minecraft.world.level.block.Block;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Set;
+
+public class ResearchStationLootTableProvider extends BlockLootSubProvider {
+    public ResearchStationLootTableProvider(HolderLookup.Provider registries) {
+        super(Set.of(), FeatureFlags.REGISTRY.allFlags(), registries);
+    }
+
+    @Override
+    protected void generate() {
+        dropSelf(ResearchStationModule.RESEARCH_STATION.get());
+    }
+
+    @Override
+    protected @NotNull Iterable<Block> getKnownBlocks() {
+        return ResearchStationModule.BLOCKS.getEntries().stream().map(Holder::value)::iterator;
+    }
+}