+++ src/main/java/dev/gacbl/logicore/data/ModBlockTagProvider.java
@@ -11,6 +11,7 @@ import dev.gacbl.logicore.blocks.datacenter.DatacenterModule;
 import dev.gacbl.logicore.blocks.datacenter_port.DatacenterPortModule;
 import dev.gacbl.logicore.blocks.drone_bay.DroneBayModule;
 import dev.gacbl.logicore.blocks.generator.GeneratorModule;
+import dev.gacbl.logicore.blocks.research_station.ResearchStationModule;
 import dev.gacbl.logicore.blocks.serverrack.ServerRackModule;
 import dev.gacbl.logicore.core.ModTags;
 import net.minecraft.core.HolderLookup;
@@ -114,6 +115,14 @@ public class ModBlockTagProvider extends BlockTagsProvider {
         tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "eclipsegate_deny"))).add(AdvanceBatteryModule.ADVANCE_BATTERY.get());
         tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "phase_deny"))).add(AdvanceBatteryModule.ADVANCE_BATTERY.get());
         tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "swapper_deny"))).add(AdvanceBatteryModule.ADVANCE_BATTERY.get());
+
+        tag(BlockTags.MINEABLE_WITH_PICKAXE).add(ResearchStationModule.RESEARCH_STATION.get());
+        tag(BlockTags.NEEDS_IRON_TOOL).add(ResearchStationModule.RESEARCH_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "eclipsegate_deny"))).add(ResearchStationModule.RESEARCH_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "phase_deny"))).add(ResearchStationModule.RESEARCH_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("justdirethings", "swapper_deny"))).add(ResearchStationModule.RESEARCH_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("c", "relocation_not_supported"))).add(ResearchStationModule.RESEARCH_STATION.get());
+        tag(BlockTags.create(ResourceLocation.fromNamespaceAndPath("mekanism", "cardboard_blacklist"))).add(ResearchStationModule.RESEARCH_STATION.get());
 
         setupOptionalTags();
     }