+++ CHANGELOG.md
@@ -30,3 +30,4 @@
     - Advance
 - Added AE2 integration via Cloud Interface block
 - Added stack upgrade item for Compiler
+- Added research station